Move components to op_primer/
