+++ app/components/op_primer/quick_filter/segmented_component/item.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# -- copyright
+#-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
 #
@@ -26,20 +26,21 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-# ++
+#++
 
-module QuickFilter
-  class BooleanComponent < SegmentedComponent
-    def initialize(name:, query:, filter_key:, true_label:, false_label:, path_args:, orders: nil)
-      super(name:, query:, filter_key:, path_args:, orders:)
+module OpPrimer
+  module QuickFilter
+    class SegmentedComponent < ApplicationComponent
+      class Item < ApplicationComponent
+        attr_reader :label, :value
 
-      @true_label = true_label
-      @false_label = false_label
-    end
+        def initialize(label:, value:)
+          super
 
-    def before_render
-      with_item(label: @true_label, value: "t")
-      with_item(label: @false_label, value: "f")
+          @label = label
+          @value = value
+        end
+      end
     end
   end
 end
